Create operations
Create operations for adding, subtracting, multiplying and dividing two
numbers.

Also created a function to get two numbers and an operator, to identify
which operation to run.

Now we need to create the HTML for the calculator itself. After that,
we should create the event listeners and DOM manipulation to make
the calculator work.